Fixed score data type

+++ src/main/java/searchengine/Score.java
@@ -1,8 +1,18 @@
 package searchengine;
 
 /**
- * Created by user on 13/11/2016.
+ *  This interface refers to an object responsible for calculating the scores of words
+ *  across a website relative to a database of websites
+ *  @author Lucas Beck
  */
 public interface Score {
-    Integer getScore (String word, Website site, Index index);
+    /**
+     * This method returns the score of a particular word given an website
+     * and a index (document)
+     * @param word word to be assigned a score
+     * @param site a website the word is contained
+     * @param index the index of the database of websites
+     * @return a float point number representing the score of the word
+     */
+    float getScore (String word, Website site, Index index);
 }
